npr_main/requirements.txt
fast-plate-ocr[onnx]==1.0.2
fastapi==0.124.2
numpy==2.2.6
onnx==1.20.0
opencv-python-headless==4.12.0.88
requests==2.32.5
ultralytics==8.3.213
uvicorn==0.38.0
Pillow
